% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinionR_data_processing.R
\name{epinion_read_labelled_xlsx}
\alias{epinion_read_labelled_xlsx}
\title{Read cleaned dictionary and apply to the data}
\usage{
epinion_read_labelled_xlsx(x, filename, dict_sheet = "dictionary")
}
\arguments{
\item{x}{datainput (dataframe).}

\item{filename}{name of the input dictionary (Excel file).}

\item{dict_sheet}{state sheet name of dictionary sheet}

\item{...}{Additional arguments may apply.}
}
\value{
A dataframe in which variable labels and value labels are cleaned
}
\description{
This function allows users to apply cleaned variable labels and value labels to
the dataframe.
}
\examples{
# Set the example folder of the package as the working directory
setwd(system.file("extdata", package = "epinionPS"))

# Read the data
mydata <- epinion_read_data(file = "sample_test_report.sav")

# Read cleaned dictionary and apply to the data
mydata <- epinion_read_labelled_xlsx(mydata, "dic - Cleaned.xlsx")
}
