% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinionR_reporting_Word.R
\name{epinion_reporting_Word}
\alias{epinion_reporting_Word}
\title{Generate Word report}
\usage{
epinion_reporting_Word(
  df,
  orderform_folder = getwd(),
  output_folder = getwd(),
  listOrderForm
)
}
\arguments{
\item{df}{datainput (dataframe).}

\item{orderform_folder}{directory where orderforms stored. Default is the current working dir.}

\item{output_folder}{directory where output report should be placed. Default is the current working dir.}

\item{listOrderForm}{list of Orderform.}

\item{...}{Additional arguments may apply.}
}
\value{
Epinion Standard Word report
}
\description{
This function allows users to generate a simple Epinion Standard Word report based on
a prepared Word template and an OrderForm.
}
\examples{
# Set the example folder of the package as the working directory
setwd(system.file("extdata", package = "epinionPS"))

# Read the data
mydata <- epinion_read_data(file = "sample_test_report.sav")

# State the default template to generate report
report_template <- "Tabular_Landscape_Template_20221116.docx"

# Get all Word OrderForms in the folder
list_orderForm <- list.files(path = paste0(getwd(), "/"), pattern='OrderForm_Word_*')

# Run reports
epinion_reporting_Word(mydata,
                       orderform_folder = paste0(getwd(), "/"),
                       output_folder = paste0(getwd(), "/"),
                       list_orderForm)
}
