% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinionR_data_processing.R
\name{epinion_write_labelled_xlsx}
\alias{epinion_write_labelled_xlsx}
\title{Save dictionary for cleaning value labels and variable labels}
\usage{
epinion_write_labelled_xlsx(
  x,
  filename,
  data_sheet = "data",
  dict_sheet = "dictionary",
  remove_repeated = FALSE,
  use_references = FALSE
)
}
\arguments{
\item{x}{datainput (dataframe).}

\item{filename}{name of the output Excel file to be saved.}

\item{data_sheet}{(optional) state sheet name if would like to export the data.}

\item{dict_sheet}{state sheet name of dictionary sheet}

\item{...}{Additional arguments may apply.}
}
\value{
Working dictionary of the datafile
}
\description{
This function allows users to generate a Excel file which includes variable labels
and value labels for reviewing and cleaning.
}
\examples{
# Set the example folder of the package as the working directory
setwd(system.file("extdata", package = "epinionPS"))

# Read the data
mydata <- epinion_read_data(file = "sample_test_report.sav")

# Save dictionary for cleaning value labels and variable labels
epinion_write_labelled_xlsx(mydata, "dic.xlsx", data_sheet = "")
}
